Use the new dispatcher and make sure everything still works
